Fix Android tribe waiver, RSVP weight, and event reminder crash.
Align waiver payload and event weight types with Supabase/iOS, clear the tab bar inset, and fall back when exact alarms are denied so Going no longer kills the app.

+++ RuckTrackerAndroid/app/src/main/java/com/artless/rucktracker/data/remote/EventRepository.kt
@@ -58,12 +58,13 @@ class EventRepository @Inject constructor(private val client: SupabaseClient) {
         client.postgrest.from("club_events").delete { filter { eq("id", eventId) } }
     }
 
-    suspend fun rsvpToEvent(eventId: String, userId: String, status: String, declaredWeight: Double?) {
+    suspend fun rsvpToEvent(eventId: String, userId: String, status: String, declaredWeight: Int?) {
         val existing = getUserRsvp(eventId, userId)
         val payload = buildJsonObject {
             put("event_id", eventId)
             put("user_id", userId)
             put("status", status)
+            // Column is integer — Doubles like 45.0 cause "invalid input syntax for type integer"
             declaredWeight?.let { put("declared_weight", it) }
         }
         if (existing != null) {
